Working settings switch
